autocode(5): remove unused variables and fix type warnings
- Remove unused variables across patient.c, doctor.c, admin.c, data_storage.c
- Fix %zu/int mismatch in admin.c schedule display
- Fix pointer-to-int-cast for srand seed in common.c
- Remove const-discarding strcpy in login.c change_password

+++ src/common.c
@@ -246,7 +246,7 @@ void generate_id(char *buffer, int buffer_size, const char *prefix) {
     static int initialized = 0;
 
     if (!initialized) {
-        srand((unsigned int)(now ^ (unsigned int)&counter));
+        srand((unsigned int)(now ^ (unsigned long long)&counter));
         counter = rand() % 7919;  // random offset to reduce restart collision
         initialized = 1;
     }
